+++ neuralprophet/forecaster.py
@@ -18,7 +18,7 @@ from neuralprophet.plot_forecast_plotly import plot as plot_plotly
 from neuralprophet.plot_forecast_plotly import plot_components as plot_components_plotly
 from neuralprophet.plot_model_parameters_matplotlib import plot_parameters
 from neuralprophet.plot_model_parameters_plotly import plot_parameters as plot_parameters_plotly
-from neuralprophet.utils_plot import log_warning_deprecation_plotly
+from neuralprophet.plot_utils import log_warning_deprecation_plotly
 
 log = logging.getLogger("NP.forecaster")
